% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdal.R
\name{gdal_warp_reproject}
\alias{gdal_warp_reproject}
\title{Warp(Reproject)}
\usage{
gdal_warp_reproject(input, output, crs)
}
\arguments{
\item{input}{raster object or <chr>. Input file source or raster object type}

\item{output}{tif. Input file destination}

\item{crs}{<chr>. Input coordinate reference system formated as 'EPSG:xxxx'}
}
\value{
RasterLayer
}
\description{
The gdalwarp utility is an image mosaicing, reprojection and warping utility. The program can
reproject to any supported projection, and can also apply GCPs stored with the image if the image is "raw"
with control information.
}
\examples{
gdal_warp_reproject(input = raster,
                    output = 'Data/Spatial/Landclass/NLCD1992/LC9240591483.tif',
                    crs = 'EPSG:4326')

gdal_warp_reproject(input = 'Data/Spatial/Landclass/NLCD1992/Original.tif',
                    output = 'Data/Spatial/Landclass/NLCD1992/Warped.tif',
                    crs = 'EPSG:4326')


}
