% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grass.R
\name{grass_slope}
\alias{grass_slope}
\title{Slope}
\usage{
grass_slope(input, output, format = 1, precision = 0, z_factor = 1,
  min_slope = 0)
}
\arguments{
\item{input}{raster object or <chr>. Input file source or raster object type}

\item{output}{tif. Input file destination}

\item{format}{Format for reporting the slope. 0 - degrees, 1 - percent}

\item{precision}{Type of output aspect and slope maps. 0 - float, 1 - double, 2 - int}

\item{z_factor}{multiplicative factor to convert elevation units to meters}

\item{min_slope}{can be used to specify the minimum slope which is computed}
}
\value{
RasterLayer
}
\description{
Generates raster maps of slope, aspect, curvatures and partial derivatives from an elevation raster map
}
\examples{
slope <- grass_slope(input = 'Spatial/DEM/DEM.tif',
                     output = 'Spatial/DEM/new.tif',
                     z_factor = 0.3048)

slope <- grass_slope(input = DEM,
                     output = 'Spatial/DEM/new.tif',
                     z_factor = 1)

}
