% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis.R
\name{qgis_clip}
\alias{qgis_clip}
\title{QGIS Clip}
\usage{
qgis_clip(input, overlay, output)
}
\arguments{
\item{input}{Target layer. Use either a string file destination or simple feature dataframe}

\item{overlay}{Object you will use to clip the target layer. Use either a string file destination or simple feature dataframe}

\item{output}{string file destination. .geojson or .shp}
}
\value{
simple feature, geojson, or shapefile
}
\description{
This algorithm clips a vector layer using the polygons of an 
additional polygons layer. Only the parts of the features in the input layer 
that falls within the polygons of the clipping layer will be added to the 
resulting layer.
}
\examples{

}
