% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis.R
\name{qgis_fixeddistancebuffer}
\alias{qgis_fixeddistancebuffer}
\title{Fixed distance buffer}
\usage{
qgis_fixeddistancebuffer(input, output, distance, segments = 5,
  dissolve = FALSE)
}
\arguments{
\item{input}{<chr> or simple feature object
Target layer. Use either a file destination or simple feature dataframe}

\item{output}{<chr>
Use a string file destination}

\item{distance}{<num>
Distance units are equivelent to coordinate system units of input shapefile/geojson/sf}

\item{segments}{<num>
Numeric value which determines how many segments are used to draw each section of the buffer. A higher value gives a smoother shape}

\item{dissolve}{<boolean>
TRUE or FALSE. TRUE dissolves all buffers into one continous polygon}
}
\value{
simple feature, geojson, or shapefile(.shp)
}
\description{
This algorithm computes a buffer area for all the features in an input layer, using a fixed distance.
}
\examples{
qgis_fixeddistancebuffer(input = 'G:/DC Policy Center/DC_One/Data/Spatial/Metro_GIS_shapefiles/Line.shp',
                         output = 'G:/DC Policy Center/DC_One/Data/Spatial/Metro_GIS_shapefiles/buffer.shp',
                         distance = 1000)

qgis_fixeddistancebuffer(input = Lines,
                         output = 'G:/DC Policy Center/DC_One/Data/Spatial/Metro_GIS_shapefiles/buffer.shp',
                         distance = 1000)


}
