% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis.R
\name{qgis_pointslayerfromtable}
\alias{qgis_pointslayerfromtable}
\title{Point Layer from Table}
\usage{
qgis_pointslayerfromtable(input, output, xfield, yfield, crs = "EPSG:4326")
}
\arguments{
\item{input}{csv file location. string}

\item{output}{string file destination. .geojson or .shp}

\item{xfield}{csv column containing X coordinate. String}

\item{yfield}{csv column containing Y coordinate. String}

\item{crs}{Coordinate reference system. Defualt is 'EPSG:4326'}
}
\value{
simple feature, geojson, or shapefile
}
\description{
Point Layer from Table
}
\examples{


}
