% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis.R
\name{qgis_polygoncentroids}
\alias{qgis_polygoncentroids}
\title{Polygon centroids}
\usage{
qgis_polygoncentroids(input, output)
}
\arguments{
\item{input}{<chr> or simple feature object
Target layer. Use either a file destination or simple feature dataframe}

\item{output}{<chr>
Destination of output file}
}
\value{
simple feature, geojson, or shapefile
}
\description{
This algorithm creates a new point layer, with points
representing the centroid of polygons of an input layer.

The attributes associated to each point in the output layer are the
 same ones associated to the original polygon.
}
\examples{

qgis_polygoncentroids(census_tracts,
                      'C://Documents/Spatial/centroid.geojson')

qgis_polygoncentroids('C://Documents/Spatial/census_tracts.geojson',
                      'C://Documents/Spatial/centroid.geojson')


}
