% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis.R
\name{qgis_countpointsinpolygon}
\alias{qgis_countpointsinpolygon}
\title{Count point in polygons}
\usage{
qgis_countpointsinpolygon(polygon, points, field = "Count", output)
}
\arguments{
\item{polygon}{Target layer. Use either a string file destination or simple feature dataframe}

\item{points}{Point layer you are counting. Use either a string file destination or simple feature dataframe}

\item{field}{Name of new atrribute column. string}

\item{output}{string file destination. .geojson or .shp}
}
\value{
simple feature, geojson, or shapefile
}
\examples{


}
