% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgis.R
\name{qgis_voronoipolygons}
\alias{qgis_voronoipolygons}
\title{Create Voronoi Polygons}
\usage{
qgis_voronoipolygons(input, output, buffer = 0)
}
\arguments{
\item{input}{<chr> or simple feature object
Input point layer. Must be desination file or simple feature}

\item{output}{<chr>
destination of output file}

\item{buffer}{<Num>
Buffer setting. Defult is 0}
}
\value{
simple feature, geojson, or shapefile
}
\description{
This algorithm takes a points layer and generates a polygon layer containing
 the voronoi polygons corresponding to those input points.
}
\examples{
qgis_voronoipolygons(input = 'C://Documents/Spatial/Points.geojson',
                               output = 'C://Documents/Spatial/Voronoi.geojson',
                               buffer = 0)

qgis_voronoipolygons(input = Points,
                     output = 'C://Documents/Spatial/Voronoi.geojson',
                     buffer = 0)



}
